chore: 글로벌 스타일 별도 지정
